+++ app/scripts/config/config.js
@@ -8,6 +8,7 @@ module.exports = {
     }
   },
   apiRoot: process.env.APIROOT || 'https://example.com',
+  kibanaRoot: process.env.KIBANAROOT || false,
   graphicsPath: (process.env.BUCKET || '') + '/graphics/',
   enableRecovery: process.env.ENABLE_RECOVERY || false
 };